Add AdminNavbar component and refactor header links; implement AdminItemManager for managing products
