+++ frontend/src/Styles/AdminNavbar.css
@@ -1,0 +1,33 @@
+.admin-nav {
+  display: flex;
+  justify-content: center;
+  gap: 2rem;
+}
+
+.admin-nav-link {
+  color: white;
+  text-decoration: none;
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  padding: 0.5rem 1rem;
+  border-radius: 0.5rem;
+  transition: background-color 0.3s;
+}
+
+.admin-nav-link:hover {
+  transform: scale(1.05);
+  /* box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2); */
+}
+
+.admin-nav-link i {
+  font-size: 1.2rem;
+}
+
+@media (max-width: 768px) {
+  .admin-nav {
+      flex-direction: column;
+      align-items: center;
+      gap: 1rem;
+  }
+}
